+++ compiler-core/src/metadata.rs
@@ -630,6 +630,18 @@ impl RemapIds {
                     .map(|element| self.type_(element))
                     .collect(),
             },
+            Type::Alias {
+                publicity,
+                aliased,
+                parameters,
+            } => Type::Alias {
+                publicity,
+                aliased: self.type_(aliased),
+                parameters: parameters
+                    .into_iter()
+                    .map(|parameter| self.type_(parameter))
+                    .collect(),
+            },
         };
         Arc::new(type_)
     }