+++ compiler-core/src/metadata/tests.rs
@@ -119,6 +119,55 @@ fn empty_module() {
 }
 
 #[test]
+fn module_with_alias() {
+    let mut module = ModuleInterface {
+        warnings: vec![],
+        is_internal: true,
+        package: "some_package".into(),
+        origin: Origin::Src,
+        name: "one/two".into(),
+        types: HashMap::new(),
+        types_value_constructors: HashMap::new(),
+        values: HashMap::new(),
+        accessors: HashMap::new(),
+        line_numbers: LineNumbers::new(""),
+        src_path: "some_path".into(),
+        minimum_required_version: Version::new(0, 1, 0),
+        type_aliases: HashMap::new(),
+        documentation: Vec::new(),
+        contains_echo: false,
+
+        references: References::default(),
+        inline_functions: HashMap::new(),
+    };
+    let _ = module.type_aliases.insert(
+        "MyAlias".into(),
+        TypeAliasConstructor {
+            publicity: Publicity::Public,
+            module: "thepackage".into(),
+            type_: Arc::new(Type::Alias {
+                aliased: Arc::new(Type::Named {
+                    publicity: Publicity::Private,
+                    package: "".into(),
+                    module: "gleam".into(),
+                    name: "Int".into(),
+                    arguments: vec![],
+                    inferred_variant: None,
+                }),
+                parameters: vec![],
+                publicity: Publicity::Public,
+            }),
+            arity: 0,
+            deprecation: Deprecation::NotDeprecated,
+            documentation: None,
+            origin: SrcSpan::default(),
+            parameters: vec![],
+        },
+    );
+    assert_eq!(roundtrip(&module), module);
+}
+
+#[test]
 fn with_line_numbers() {
     let module = ModuleInterface {
         warnings: vec![],